+++ pkg/archtest/acceptance_tests_arch_check_test.go
@@ -1,4 +1,4 @@
-package archtests
+package archtest
 
 import (
 	"testing"
