+++ pkg/archtest/file_filter_providers.go
@@ -1,4 +1,4 @@
-package archtests
+package archtest
 
 import "regexp"
